fix(app): stop a microphone-only recording reporting phantom app-track problems
Three places still spoke as if every recording had an app track.

`buildRecording` logged the negotiated app format, and with no tap the rate and
channel count fall back to the *configured* values, so every microphone-only
recording blamed a mono USB device and a renegotiated rate for hardware nothing
touched. That lands in the log a support bundle is read from. It also opened the
mic WAV a second time to compute a duration for a rate cross-check that cannot
apply without an app track.

`ChannelHealthController` never reset its topology in `stop()` and set it after
both guards in `start()`, so a start that turned back early left the previous
recording deciding what the icon painted. The overlay getters also read the
topology first, and `&&` short-circuits, so on a recording without that channel
the observable flags were never read and registered no dependency for that
render pass; the flags are read into a local first now.

`CapturedChannels` could express "neither channel", which is the state
`RecordingSource` exists to rule out. Its memberwise init is private and the
three named values mirror the three source cases, so a projection cannot hand
back what the thing it projects makes unrepresentable.

`buildRecording` and its helpers moved to their own file, along the seam their
own doc comment already described: pure file processing, no capture session, no
availability gate. That is what put the parent file back under the line cap.
